FLASK_APP/requirements.txt
flask
python-dotenv
mysql-connector-python 
PyJWT